pyproject.toml
[project]
name = "fastapi-project-template"
version = "0.1.0"
description = "A template for a FastAPI project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { file = "LICENSE" }
dependencies = [
  "bcrypt==4.0.1",
  "fastapi[standard]>=0.115.12",
  "httpx>=0.28.1",
  "loguru>=0.7.3",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.9",
  "pydantic>=2.11.4",
  "pydantic-settings[toml]>=2.9.1",
  "pyjwt>=2.10.1",
  "sqlmodel>=0.0.24",
  "typer>=0.15.4",
]

[dependency-groups]
dev = ["git-cliff>=2.8.0", "gitpython>=3.1.44", "packaging>=25.0", "pre-commit>=4.2.0", "ruff>=0.11.9"]

[project.scripts]
# Release tool: Update version number, generate changelog, and create git tag.
release = "scripts.release:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
